requirements.txt
Flask==1.1.2
googlemaps==4.4.2
python-dotenv==0.15.0
wikipedia==1.4.0
